rename weight chart functions
